// src/app/_core/models/models.ts
export interface CriterionValueVO {
  id: number;
  value: string;
  isActive: boolean;
  criterionId: number;
  criterionLabel: string;
}

export interface CustomerFullVO {
  id: number;
  customer: CustomerVO;
  person: PersonVO;
  personTit: PersonVO;
  holderContracts: CustomerVO[];
}

export interface CustomerHardwareParkItemVO {
  id: number;
  status: string;
  subscriptionDate: Date;
  contactMethodPostalAddressId: number;
  delimitedContactMethodPostalAddress: string;
  createdAt: Date;
  createdById: number;
  createdByFullName: string;
  updatedAt: Date;
  updatedById: number;
  updatedByFullName: string;
  customerId: number;
  terminatedAt: Date;
  terminatedById: number;
  terminatedByFullName: string;
  comments: string;
  productLabel: string;
  associatedLineId: number;
  associatedLineIdentifier: string;
  productId: number;
  product: ProductVO;
  productFamily: FamilyVO;
  facturationLine: FacturationLineVO;
  billAccountIdentifier: string;
  recurrent: boolean;
  toInvoice: boolean;
  alreadyPaidInvoice: boolean;
  unreferencedProductCategory: FamilyVO;
  serial: string;
  masterProductName: string;
  productCategoryName: string;
  productCategoryKey: string;
  productFamilyName: string;
  productSubFamilyName: string;
  productQuantityBlocked: boolean;
  productSerialRequired: boolean;
  productDiscountIncluded: boolean;
  nbOffersAvailable: number;
  hasOffers: boolean;
  maxNbOffers: number;
  customerName: string;
  requestType: string;
  requestId: number;
  billLivrableURL: string;
  billFileName: string;
  billResolution: boolean;
  warrantyLabel: string;
  warrantyValue: number;
  articleClassId: number;
  manufacturerId: number;
  acquisitionPriceReal: number;
  nomenclature: NomenclatureVO;
  customerBeneficeryId: number;
  type: string;
  category: string;
  customerNicheIdentifier: string;
  offerLabel: string;
  customerBeneficeryName: string;
  billAccountNicheIdentifier: string;
  hardwareStatus : string; 
}
// ...
export interface FacturationLineVO {
  label: string;
  unitPrice: number;
  tva: number;
  quantity: number;
  remise: number;
}

export interface FamilyVO {
  id: number;
  name: string;
  parent: FamilyVO;
  categoryKey: string;
}
// ...
export interface MasterProductVO {
  id: number;
  name: string;
  partnerId: number;
  partnerName: string;
  longDescription: string;
  shortDescription: string;
  stockQuantity: number;
  stockOrder: number;
  isOnStock: boolean;
  manufacturerId: number;
  manufacturerName: string;
  availabilityDate: Date;
  partnerProductReference: string;
  proratedAtBegin: boolean;
  proratedAtEnd: boolean;
  recurrent: boolean;
  subscriptionPeriodicity: string;
  productImageName: string;
  productDescriptionName: string;
  billingCategory: string;
  unitOfWorkId: number;
  unitOfWorkLabel: string;
  interventionDuration: number;
  detail: string;
  prerequisite: string;
  pricingConditions: string;
  quantityBlocked: boolean;
  commentMandatory: boolean;
  serialRequired: boolean;
  needs: NeedVO[];
  features: CriterionValueVO[];
  serialNumberPrefix: ReferenceDataVO;
  warrantyValue: number;
  productsVO: AdminProductVO[];
  acquisitionPrice: number;
  acquisitionPriceReal: number;
  articleClassId: number;
  nomenclature: NomenclatureVO;
  refDataArticleClass: ReferenceDataVO;
  annualReport: string;
  travelIncluded: boolean;
  pricingType: string;
  pricingTypeReal: string;
  marginRate: number;
  marginRateReal: number;
  marginValue: number;
  marginValueReal: number;
  orangeProductReference: string;
  publidispatchStockQuantity: number;
  isOnPublidispatchStock: boolean;
  ismodified: boolean;
  eanCode: string;
}
// ...
export interface NeedVO {
  id: number;
  label: string;
  parent: NeedVO;
  childs: NeedVO[];
}
// ...
export interface NomenclatureVO {
  id: number;
  value: string;
  label: string;
  level: number;
  parent: NomenclatureVO;
  childrenIds: number[];
}
// ...
export interface PenicheBillAccountLightVO extends Comparable<PenicheBillAccountLightVO> {
  billAccountUniverse: string;
  billAccountIdentifier: string;
  billAccountStatus: string;
  billAccountBalance: number;
  billAccountName: string;
  billAccountAlias: string;
  billAccountRecurrent: boolean;
  billAccountBillingDay: number;
  billAccountPaymentMode: string;
  billAccountPayerTitle: string;
  billAccountPayerLastName: string;
  billAccountPayerFirstName: string;
  billAccountPayerCompany: string;
  billAccountTitularTitle: string;
  billAccountTitularLastName: string;
  billAccountTitularFirstName: string;
  billAccountTitularCompany: string;
  billAccountTitularAddress: string;
  billAccountPayerAddress: string;
  billAccountTvaExemption: boolean;
  billAccountLines: PenicheBillAccountLineVO[];
  billAccountNicheIdentifiant: string;
  billAccountRecurrentLabel: string;
  paymentModeLabel: string;
  payerName: string;
}

export interface PenicheBillAccountLineVO {
  lineIdentifier: string;
  lineStatus: string;
  lineDateStatus: string;
  lineAlias: string;
  active: boolean;
}
// ...
export interface ProductVO {
  id: number;
  name: string;
  priceHt: number;
  status: string;
  family: FamilyVO;
  partnerId: number;
  partnerName: string;
  longDescription: string;
  shortDescription: string;
  stockQuantity: number;
  stockOrder: number;
  isOnStock: boolean;
  manufacturerId: number;
  manufacturerName: string;
  availabilityDate: Date;
  partnerProductReference: string;
  prorataBegin: boolean;
  prorataEnd: boolean;
  recurrent: boolean;
  blocked: boolean;
  subscriptionPeriodicity: string;
  productImageName: string;
  productDescriptionName: string;
  tva: string;
  billingCategory: string;
  ecoTaxe: number;
  libelleFacture: string;
  uniteOeuvreId: number;
  uniteOeuvreLabel: string;
  prixSurDemande: boolean;
  dureeIntervention: number;
  detail: string;
  preRequis: string;
  conditionsTarifaires: string;
  showingPortal: boolean;
  quantityBlocked: boolean;
  serialRequired: boolean;
  discountLabel: string;
  discountIncluded: boolean;
  discountFree: boolean;
  included: boolean;
  discountHt: number;
  discountPercent: number;
  needs: NeedVO[];
  features: CriterionValueVO[];
  prefixSerialNumber: ReferenceDataVO;
  warrantyValue: number;
  warrantyLabel: string;
  isWarranty: boolean;
  articleClassLabel: string;
  articleClassId: number;
  nomenclature: NomenclatureVO;
  commentMandatory: boolean;
  masterProductId: number;
  masterProduct: MasterProductVO;
  customerTarget: string;
  acquisitionPrice: number;
  acquisitionPriceReal: number;
  purchasePrice: number;
  purchasePriceReal: number;
  pricingType: string;
  pricingTypeReal: string;
  travelIncluded: boolean;
  fixedMarginRate: number;
  fixedMarginRateReal: number;
  publidispatchStockQuantity: number;
  isOnPublidispatchStock: boolean;
  stockOpera: number;
  stockRiviera: number;
  stockRaspail: number;
  stockProvence: number;
  stockLyonRhone: number;
  stockHauteSavoie: number;
  stockLille : number;
  available: string;
  photo: any;
}

export interface PurchaseRequestVO extends RequestVO {
  refRequestOrigin: ReferenceDataVO;
}

export interface ReferenceDataVO {
  id: number;
  label: string;
  ordinal: number;
  active: boolean;
  dateUpdate: Date;
  parent: ReferenceDataVO;
  list: ReferenceListVO;
  children: ReferenceDataVO[];
  logoName: string;
  labelComplement: string;
  key: string;
}
// ...
export interface Comparable<T> {
}